requirements.txt
grpcio==1.59.0
grpcio-tools==1.58.0
requests==2.32.2